convert skybox
