+++ Engine/src/component/Component.cpp
@@ -18,14 +18,7 @@ void TagComponent::attachToEntity(std::shared_ptr<Component> c, Entity entityHan
     attachSimple<TagComponent>(c, entityHandler);
 }
 
-void SkyboxComponent::attachToEntity(std::shared_ptr<Component> c, Entity entityHandler, Scene& scene)
-{
-	if (auto sc = std::dynamic_pointer_cast<SkyboxComponent>(c))
-	{
-		auto& skyboxComponent = entityHandler.addComponent<SkyboxComponent>(*sc);
-		skyboxComponent.build();
-	}
-}
+
 
 void TestComp::attachToEntity(std::shared_ptr<Component> c, Entity entityHandler, Scene& scene)
 {
@@ -124,26 +117,3 @@ void InstanceBatch::build()
 	glVertexAttribDivisor(9, 1);
 }
 
-SkyboxComponent::SkyboxComponent(Resource<Texture> skyboxImage)
-{
-	setSkybox(skyboxImage);
-}
-
-void SkyboxComponent::setSkybox(Resource<Texture> image)
-{
-	originalImage = image;
-}
-
-void SkyboxComponent::build()
-{
-
-	// TODO check if orig image is cube and support cubemap load
-
-	cubemap = EquirectangularToCubemapConverter::fromEquirectangularToCubemap(originalImage);
-
-	auto scene = Engine::get()->getContext()->getActiveScene().get();
-	auto irradianceMap = IBL::generateIrradianceMap(cubemap, scene);
-	auto prefilterEnvMap = IBL::generatePrefilterEnvMap(cubemap, scene);
-
-	scene->setIBLData(irradianceMap, prefilterEnvMap);
-}
