+++ Engine/src/component/Component.h
@@ -75,31 +75,6 @@ struct EngineAPI TagComponent : public Component
         std::string tag;
 };
 
-
-
-
-struct EngineAPI SkyboxComponent : public Component
-{
-	SkyboxComponent() = default;
-
-	SkyboxComponent(Resource<Texture> skyboxImage);
-
-	template <class Archive>
-	void serialize(Archive& archive) {
-		SERIALIZED_MEMBER(originalImage);
-
-	}
-
-	void setSkybox(Resource<Texture> image);
-
-	void build();
-	
-
-        Resource<Texture> originalImage;
-        Resource<Texture> cubemap;
-        static void attachToEntity(std::shared_ptr<Component>, Entity, Scene&);
-};
-
 struct EngineAPI InstanceBatch : public Component
 {
 public:
