+++ Engine/src/serialize/Archiver.h
@@ -84,7 +84,7 @@ struct SerializedEntity
 	std::optional<DirectionalLight> dLight;
 	std::optional<PointLight> pLight;
 	//std::optional<ObjectComponent> obj;
-	std::optional<SkyboxComponent> skybox;
+	//std::optional<SkyboxComponent> skybox;
 	std::optional<ImageComponent> image;
     std::optional<Animator> animator;
     std::optional<Terrain> terrain;
@@ -108,7 +108,7 @@ struct SerializedEntity
 		SERIALIZED_MEMBER(dLight);
 		SERIALIZED_MEMBER(pLight);
 		//SERIALIZED_MEMBER(obj);
-		SERIALIZED_MEMBER(skybox);
+		//SERIALIZED_MEMBER(skybox);
 		SERIALIZED_MEMBER(image);
         SERIALIZED_MEMBER(animator);
         SERIALIZED_MEMBER(terrain);
